+++ R/tip_with_binary.R
@@ -1,15 +1,23 @@
-#' Tip a result with a normally distributed confounder.
+#' Tip a result with a binary confounder.
 #'
-#' choose one of the following, and the other will be estimated:
-#' * `smd`
+#' @description
+#' Choose two of the following three to specify, and the third will be estimated:
+#' * `exposed_p`
+#' * `unexposed_p`
 #' * `outcome_association`
 #'
+#' Alternatively, specify all three and the function will return the number of unmeasured
+#' confounders specified needed to tip the analysis.
+#'
+#' @details [`tip_b()`] is an alias for [`tip_with_binary()`].
 #' @param effect Numeric positive value. Observed exposure - outcome effect
 #'    (assumed to be the exponentiated coefficient, so a relative risk, odds
 #'    ratio, or hazard ratio). This can be the point estimate, lower confidence
 #'    bound, or upper confidence bound.
-#' @param smd Numeric. Estimated difference in scaled means between the
-#'    unmeasured confounder in the exposed population and unexposed population
+#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the exposed population
+#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the unexposed population
 #' @param outcome_association Numeric positive value. Estimated association
 #'    between the unmeasured confounder and the outcome
 #' @param verbose Logical. Indicates whether to print informative message.
@@ -22,17 +30,17 @@
 #'   this to "or"; if you are supplying a risk ratio or your outcome is rare,
 #'   set this to "none" (default).
 #'
-#'
-#' @return Data frame.
-#'
 #' @examples
 #' ## to estimate the association between an unmeasured confounder and outcome
 #' ## needed to tip analysis
-#' tip(1.2, smd = -2)
+#' tip_with_binary(1.2, exposed_p = 0.5, unexposed_p = 0)
 #'
 #' ## to estimate the number of unmeasured confounders specified needed to tip
 #' ## the analysis
-#' tip(1.2, smd = -2, outcome_association = .99)
+#' tip_with_binary(1.2,
+#'   exposed_p = 0.5,
+#'   unexposed_p = 0,
+#'   outcome_association = 1.1)
 #'
 #' ## Example with broom
 #' if (requireNamespace("broom", quietly = TRUE) &&
@@ -41,27 +49,38 @@
 #'    broom::tidy(conf.int = TRUE, exponentiate = TRUE) %>%
 #'    dplyr::filter(term == "mpg") %>%
 #'    dplyr::pull(conf.low) %>%
-#'    tip(outcome_association = 2.5)
+#'    tip_with_binary(exposed_p = 1, outcome_association = 1.15)
 #'}
 #' @export
-tip <- function(effect, smd = NULL, outcome_association = NULL, verbose = TRUE,
-                correction_factor = "none") {
-    o <- purrr::map(
-      effect,
-      ~ tip_one(
-        .x,
-        smd = smd,
-        outcome_association = outcome_association,
-        verbose = verbose,
-        correction_factor
-      )
-    )
-    do.call(rbind, o)
-  }
+tip_with_binary <- function(effect,
+                            exposed_p = NULL,
+                            unexposed_p = NULL,
+                            outcome_association = NULL,
+                            verbose = TRUE,
+                            correction_factor = "none") {
 
-tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
-  check_effect(b)
-  check_gamma(outcome_association)
+  o <- purrr::map(
+    effect,
+    ~ tip_with_binary_one(
+      .x,
+      exposed_p = exposed_p,
+      unexposed_p = unexposed_p,
+      outcome_association = outcome_association,
+      verbose = verbose,
+      correction_factor = correction_factor
+    )
+  )
+  do.call(rbind, o)
+}
+
+tip_with_binary_one <- function(b,
+                                exposed_p,
+                                unexposed_p,
+                                outcome_association,
+                                verbose,
+                                correction_factor) {
+
+  n_unmeasured_confounders <- 1
 
   correction <- ""
   if (correction_factor == "hr") {
@@ -76,20 +95,26 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
     correction <- 'You opted to use the odds ratio correction to convert your odds ratios to approximate risk ratios.\nThis is a good idea if the outcome is common (>15%).'
   }
 
-  n_unmeasured_confounders <- 1
-
   if (is.null(outcome_association)) {
-    outcome_association <- b ^ (1 / smd)
-  } else if (is.null(smd)) {
-    smd <- log(b) / log(outcome_association)
+    outcome_association <- tip_gamma(unexposed_p, exposed_p, b)
+  } else if (is.null(unexposed_p)) {
+    unexposed_p <- tip_p0(exposed_p, outcome_association, b)
+  } else if (is.null(exposed_p)) {
+    exposed_p <- tip_p1(unexposed_p, outcome_association, b)
   } else {
     n_unmeasured_confounders <-
-      log(b) / (smd * log(outcome_association))
+      tip_n(unexposed_p, exposed_p, outcome_association, b)
+
     if (any(n_unmeasured_confounders < 0)) {
-      if (length(smd) > 1) {
-        smds <- smd[n_unmeasured_confounders < 0]
+      if (length(unexposed_p) > 1) {
+        unexposed_ps <- unexposed_p[n_unmeasured_confounders < 0]
       } else {
-        smds <- smd
+        unexposed_ps <- unexposed_p
+      }
+      if (length(exposed_p) > 1) {
+        exposed_ps <- exposed_p[n_unmeasured_confounders < 0]
+      } else {
+        exposed_ps <- exposed_p
       }
       if (length(outcome_association) > 1) {
         outcome_associations <- outcome_association[n_unmeasured_confounders < 0]
@@ -99,7 +124,8 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
 
       warning_glue(
         "The observed effect {b} would not tip with the unmeasured confounder given:",
-        "\n  * `smd`: {smds}",
+        "\n  * `exposed_p`: {exposed_ps}",
+        "\n  * `unexposed_p`: {unexposed_ps}",
         "\n  * `outcome_association`: {outcome_associations}\n\n"
       )
       n_unmeasured_confounders <- max(0, n_unmeasured_confounders)
@@ -107,34 +133,47 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
     too_small <-
       n_unmeasured_confounders < 1 & n_unmeasured_confounders > 0
     if (any(too_small)) {
-      smds <- ifelse(length(smd) > 1, smd[too_small], smd)
-      outcome_associations <-
-        ifelse(length(outcome_association) > 1,
-               outcome_association[too_small],
-               outcome_association)
+      if (length(unexposed_p) > 1) {
+        unexposed_ps <- unexposed_p[too_small]
+      } else {
+        unexposed_ps <- unexposed_p
+      }
+      if (length(exposed_p) > 1) {
+        exposed_ps <- exposed_p[too_small]
+      } else {
+        exposed_ps <- exposed_p
+      }
+      if (length(outcome_association) > 1) {
+        outcome_associations <- outcome_association[too_small]
+      } else {
+        outcome_associations <- outcome_association
+      }
       warning_glue(
         "The observed effect {b} would tip with < 1 of the given unmeasured confounders:",
-        "\n  * `smd`: {smds}",
+        "\n  * `exposed_p`: {exposed_ps}",
+        "\n  * `unexposed_p`: {unexposed_ps}",
         "\n  * `outcome_association`: {outcome_associations}\n\n"
       )
     }
   }
   o <- tibble::tibble(
     observed_effect = b,
-    smd = smd,
+    exposed_p = exposed_p,
+    unexposed_p = unexposed_p,
     outcome_association = outcome_association,
     n_unmeasured_confounders = n_unmeasured_confounders
   )
   if (verbose) {
     if (all(o$n_unmeasured_confounders == 0)) {
       o_notip <- o[o$n_unmeasured_confounders == 0,]
+
       message_glue(
         "The observed effect ({round(o_notip$observed_effect, 2)}) ",
         "cannot be tipped by an unmeasured confounder\nwith the ",
         "following specifications:",
-        "\n  * estimated difference in scaled means between the ",
-        "unmeasured confounder\n     in the exposed population and ",
-        "unexposed population: {round(o_notip$smd, 2)}",
+        "\n  * estimated prevalence of the unmeasured confounder ",
+        "in the exposed population: {round(o_notip$exposed_p, 2)}\n  * estimated prevalence of ",
+        "the unmeasured confounder in the unexposed population: {round(o_notip$unexposed_p, 2)}",
         "\n  * estimated association between the unmeasured confounder and the ",
         "outcome: {round(o_notip$outcome_association, 2)}\n\n{correction}"
       )
@@ -144,9 +183,9 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
         "The observed effect ({round(o_notip$observed_effect, 2)}) ",
         "cannot be tipped by an unmeasured confounder\nwith the ",
         "following specifications:",
-        "\n  * estimated difference in scaled means between the ",
-        "unmeasured confounder\n     in the exposed population and ",
-        "unexposed population: {round(o_notip$smd, 2)}",
+        "\n  * estimated prevalence of the unmeasured confounder ",
+        "in the exposed population: {round(o_notip$exposed_p, 2)}\n  * estimated prevalence of ",
+        "the unmeasured confounder in the unexposed population: {round(o_notip$unexposed_p, 2)}",
         "\n  * estimated association between the unmeasured confounder and the ",
         "outcome: {round(o_notip$outcome_association, 2)}\n\n{correction}"
       )
@@ -154,12 +193,12 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
       o_tip <- o[o$n_unmeasured_confounders != 0,]
       message_glue(
         "The observed effect ({round(o_tip$observed_effect, 2)}) WOULD ",
-        "be tipped by {round(o$n_unmeasured_confounders)} ",
+        "be tipped by {round(o_tip$n_unmeasured_confounders)} ",
         "unmeasured confounder{ifelse(o_tip$n_unmeasured_confounders > 1, 's', '')}\n",
-        "with the following specifications:",
-        "\n  * estimated difference in scaled means between the ",
-        "unmeasured confounder\n    in the exposed population and ",
-        "unexposed population: {round(o_tip$smd, 2)}",
+        "with the following ",
+        "specifications:\n  * estimated prevalence of the unmeasured confounder ",
+        "in the exposed population: {round(o_tip$exposed_p, 2)}\n  * estimated prevalence of ",
+        "the unmeasured confounder in the unexposed population: {round(o_tip$unexposed_p, 2)}",
         "\n  * estimated association between the unmeasured confounder and the ",
         "outcome: {round(o_tip$outcome_association, 2)}\n\n{correction}"
       )
@@ -168,10 +207,10 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
         "The observed effect ({round(o$observed_effect, 2)}) WOULD ",
         "be tipped by {round(o$n_unmeasured_confounders)} ",
         "unmeasured confounder{ifelse(o$n_unmeasured_confounders > 1, 's', '')}\n",
-        "with the following specifications:",
-        "\n  * estimated difference in scaled means between the ",
-        "unmeasured confounder\n    in the exposed population and ",
-        "unexposed population: {round(o$smd, 2)}",
+        "with the following ",
+        "specifications:\n  * estimated prevalence of the unmeasured confounder ",
+        "in the exposed population: {round(o$exposed_p, 2)}\n  * estimated prevalence of ",
+        "the unmeasured confounder in the unexposed population: {round(o$unexposed_p, 2)}",
         "\n  * estimated association between the unmeasured confounder and the ",
         "outcome: {round(o$outcome_association, 2)}\n\n{correction}"
       )
@@ -180,51 +219,49 @@ tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
   o
 }
 
-#' Tip an observed relative risk with a normally distributed confounder.
+#' Tip an observed relative risk with a binary confounder.
 #'
-#' choose one of the following, and the other will be estimated:
-#' * `smd`
+#' @description
+#' Choose two of the following three to specify, and the third will be estimated:
+#' * `exposed_p`
+#' * `unexposed_p`
 #' * `outcome_association`
 #'
+#' Alternatively, specify all three and the function will return the number of unmeasured
+#' confounders specified needed to tip the analysis.
 #' @param rr Numeric positive value. Observed exposure - outcome relative risk.
 #'    This can be the point estimate, lower confidence bound, or upper
 #'    confidence bound.
-#' @param smd Numeric. Estimated difference in scaled means between the
-#'    unmeasured confounder in the exposed population and unexposed population
+#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the exposed population
+#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the unexposed population
 #' @param outcome_association Numeric positive value. Estimated association
 #'    between the unmeasured confounder and the outcome
 #' @param verbose Logical. Indicates whether to print informative message.
 #'    Default: `TRUE`
-#'
-#'
-#' @return Data frame.
-#'
-#' @examples
-#' ## to estimate the association between an unmeasured confounder and outcome
-#' ## needed to tip analysis
-#' tip_rr(1.2, smd = -2)
-#'
-#' ## to estimate the number of unmeasured confounders specified needed to tip
-#' ## the analysis
-#' tip_rr(1.2, smd = -2, outcome_association = .99)
-#'
-#' @export
-tip_rr <- function(rr, smd = NULL, outcome_association = NULL, verbose = TRUE) {
-  tip(rr, smd = smd, outcome_association = outcome_association, verbose = verbose)
+
+tip_rr_with_binary <- function(rr, exposed_p, unexposed_p, outcome_association, verbose = TRUE) {
+  tip_with_binary(rr, exposed_p = exposed_p, unexposed_p = unexposed_p, outcome_association = outcome_association, verbose = verbose)
 }
 
-
-#' Tip an observed hazard ratio with a normally distributed confounder.
+#' Tip an observed hazard ratio with a binary confounder.
 #'
-#' choose one of the following, and the other will be estimated:
-#' * `smd`
+#' @description
+#' Choose two of the following three to specify, and the third will be estimated:
+#' * `exposed_p`
+#' * `unexposed_p`
 #' * `outcome_association`
 #'
+#' Alternatively, specify all three and the function will return the number of unmeasured
+#' confounders specified needed to tip the analysis.
 #' @param hr Numeric positive value. Observed exposure - outcome hazard ratio.
 #'    This can be the point estimate, lower confidence bound, or upper
 #'    confidence bound.
-#' @param smd Numeric. Estimated difference in scaled means between the
-#'    unmeasured confounder in the exposed population and unexposed population
+#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the exposed population
+#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the unexposed population
 #' @param outcome_association Numeric positive value. Estimated association
 #'    between the unmeasured confounder and the outcome
 #' @param verbose Logical. Indicates whether to print informative message.
@@ -236,33 +273,33 @@ tip_rr <- function(rr, smd = NULL, outcome_association = NULL, verbose = TRUE) {
 #'    set this to `TRUE`. Default: `FALSE`.
 #'
 #' @return Data frame.
+#' @export
 #'
 #' @examples
-#' ## to estimate the association between an unmeasured confounder and outcome
-#' ## needed to tip analysis
-#' tip_hr(1.2, smd = -2)
-#'
-#' ## to estimate the number of unmeasured confounders specified needed to tip
-#' ## the analysis
-#' tip_hr(1.2, smd = -2, outcome_association = .99)
-#'
-#' @export
-tip_hr <- function(hr, smd = NULL, outcome_association = NULL, verbose = TRUE, hr_correction = FALSE) {
+#' tip_hr_with_binary(0.9, 0.9, 0.1)
+
+tip_hr_with_binary <- function(hr, exposed_p = NULL, unexposed_p = NULL, outcome_association = NULL, verbose = TRUE, hr_correction = FALSE) {
   correction_factor <- ifelse(hr_correction, "hr", "none")
-  tip(hr, smd = smd, outcome_association = outcome_association, verbose = verbose, correction_factor = correction_factor)
+  tip_with_binary(hr, exposed_p = exposed_p, unexposed_p = unexposed_p, outcome_association = outcome_association, verbose = verbose, correction_factor = correction_factor)
 }
 
-#' Tip an observed odds ratio with a normally distributed confounder.
+#' Tip an observed odds ratio with a binary confounder.
 #'
-#' choose one of the following, and the other will be estimated:
-#' * `smd`
+#' @description
+#' Choose two of the following three to specify, and the third will be estimated:
+#' * `exposed_p`
+#' * `unexposed_p`
 #' * `outcome_association`
 #'
+#' Alternatively, specify all three and the function will return the number of unmeasured
+#' confounders specified needed to tip the analysis.
 #' @param or Numeric positive value. Observed exposure - outcome odds ratio.
 #'    This can be the point estimate, lower confidence bound, or upper
 #'    confidence bound.
-#' @param smd Numeric. Estimated difference in scaled means between the
-#'    unmeasured confounder in the exposed population and unexposed population
+#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the exposed population
+#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
+#'    unmeasured confounder in the unexposed population
 #' @param outcome_association Numeric positive value. Estimated association
 #'    between the unmeasured confounder and the outcome
 #' @param verbose Logical. Indicates whether to print informative message.
@@ -274,36 +311,18 @@ tip_hr <- function(hr, smd = NULL, outcome_association = NULL, verbose = TRUE, h
 #'    set this to `TRUE`. Default: `FALSE`.
 #'
 #' @return Data frame.
+#' @export
 #'
 #' @examples
-#' ## to estimate the association between an unmeasured confounder and outcome
-#' ## needed to tip analysis
-#' tip_or(1.2, smd = -2)
-#'
-#' ## to estimate the number of unmeasured confounders specified needed to tip
-#' ## the analysis
-#' tip_or(1.2, smd = -2, outcome_association = .99)
-#'
-#' ## Example with broom
-#' if (requireNamespace("broom", quietly = TRUE) &&
-#'     requireNamespace("dplyr", quietly = TRUE)) {
-#'   glm(am ~ mpg, data = mtcars, family = "binomial") %>%
-#'    broom::tidy(conf.int = TRUE, exponentiate = TRUE) %>%
-#'    dplyr::filter(term == "mpg") %>%
-#'    dplyr::pull(conf.low) %>%
-#'    tip_or(outcome_association = 2.5, or_correction = TRUE)
-#'}
-#' @export
-tip_or <- function(or, smd = NULL, outcome_association = NULL, verbose = TRUE, or_correction = FALSE) {
+#' tip_or_with_binary(0.9, 0.9, 0.1)
+
+tip_or_with_binary <- function(or, exposed_p = NULL, unexposed_p = NULL, outcome_association = NULL, verbose = TRUE, or_correction = FALSE) {
   correction_factor <- ifelse(or_correction, "or", "none")
-  tip(or, smd = smd, outcome_association = outcome_association, verbose = verbose, correction_factor = correction_factor)
+  tip_with_binary(or, exposed_p = exposed_p, unexposed_p = unexposed_p, outcome_association = outcome_association, verbose = verbose, correction_factor = correction_factor)
 }
 
-
-#' @rdname tip
+#' @rdname tip_with_binary
 #' @export
-tip_with_continuous <- tip
+tip_b <- tip_with_binary
 
-#' @rdname tip
-#' @export
-tip_c <- tip
+